refactor(ui): moved settings impl to general component
